content_loading/requirements.txt
openai==1.56.0
azure-functions==1.21.3
azure.storage.blob==12.24.0
azure.identity==1.19.0
azure.ai.textanalytics==5.3.0
azure-cosmos==4.9.0
llama-index==0.12.0
llama-index-embeddings-azure-openai==0.3.0
llama-index-llms-azure-openai==0.3.0
llama-index-readers-azstorage_blob==0.3.0
azure-ai-textanalytics==5.3.0
presidio-analyzer
llama-index-multi-modal-llms-azure-openai==0.3.0
